feat: 假期设置更改为小数

+++ Server.Application/Entities/ELeaveRemaining.cs
@@ -6,5 +6,5 @@ public class ELeaveRemaining : BasicEntity
 {
     public string UserId { get; set; } = string.Empty;
     public string LeaveType { get; set; } = string.Empty;
-    public int Cnt { get; set; }
+    public double Cnt { get; set; }
 }
